# docs/openapi/event-service.openapi.yml
openapi: 3.0.3
info:
  title: BoardHub Event Service API
  version: 0.20.0
  description: API REST dell'event-service BoardHub per controllo dei tavoli, sessioni, ingresso dei giocatori, personaggi, eventi e movimento su griglia.
servers:
  - url: http://localhost:8082
    description: Ambiente locale di sviluppo
...
paths:
...
  /api/v1/public/sessions/{sessionId}/join-requests:
    post:
      tags:
        - Join
      summary: Richiede al DM di entrare nella sessione
      description: La stessa chiave di idempotenza e lo stesso body restituiscono la richiesta gia creata.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          example: session-20260705-001
        - name: Idempotency-Key
          in: header
          required: true
          description: UUID nuovo generato dal client per questa operazione.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJoinRequestRequest"
      responses:
        "201":
          description: Richiesta inserita nella coda del DM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequestResponse"
        "409":
          description: Sessione chiusa, richiesta duplicata o chiave riusata con altri dati
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Troppe richieste dallo stesso giocatore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/player/sessions/{sessionId}/me:
    get:
      tags:
        - Join
      summary: Restituisce l'identita attiva del giocatore autenticato
      description: >
        Verifica firma, versione e sessione della credenziale. La richiesta viene
        rifiutata anche quando il partecipante ha lasciato la sessione o la sessione
        e stata conclusa.
      security:
        - PlayerBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Partecipante autenticato e attivo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantResponse"
        "401":
          description: Credenziale assente, alterata, fuori sessione o revocata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/player/sessions/{sessionId}/characters:
    post:
      tags:
        - Characters
      summary: Crea un personaggio per il giocatore autenticato
      description: >
        Il proprietario viene ricavato dal token Bearer. Il backend applica
        validazione, limite per partecipante e valore iniziale di hpCurrent
        uguale a hpMax quando il campo non viene inviato.
      security:
        - PlayerBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterRequest"
      responses:
        "201":
          description: Personaggio creato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          description: Dati del personaggio non validi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Credenziale giocatore non valida o sessione non attiva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Limite di personaggi del partecipante raggiunto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - Characters
      summary: Elenca i personaggi posseduti dal giocatore autenticato
      description: Non espone i personaggi appartenenti agli altri partecipanti.
      security:
        - PlayerBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Personaggi del proprietario ordinati per creazione
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CharacterResponse"
        "401":
          description: Credenziale giocatore non valida o sessione non attiva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/dm/sessions/{sessionId}/join-requests:
    get:
      tags:
        - Join
      summary: Elenca le richieste di ingresso della sessione
      security:
        - DmSessionBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, ACCEPTED, REJECTED, EXPIRED]
            default: PENDING
      responses:
        "200":
          description: Richieste ordinate per data di invio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoinRequestResponse"
        "401":
          description: Credenziale DM non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/dm/sessions/{sessionId}/join-requests/{requestId}/accept:
    post:
      tags:
        - Join
      summary: Accetta una richiesta e crea il partecipante
      security:
        - DmSessionBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Partecipante attivo e credenziale locale firmata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptJoinRequestResponse"
        "401":
          description: Credenziale DM non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Richiesta non piu pendente o sessione al completo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/dm/sessions/{sessionId}/join-requests/{requestId}/reject:
    post:
      tags:
        - Join
      summary: Rifiuta una richiesta di ingresso
      security:
        - DmSessionBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Richiesta rifiutata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequestResponse"
        "401":
          description: Credenziale DM non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/dm/sessions/{sessionId}/participants:
    get:
      tags:
        - Join
      summary: Elenca i partecipanti attivi della sessione
      security:
        - DmSessionBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Partecipanti attivi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParticipantResponse"
  /api/v1/dm/sessions/{sessionId}/characters:
    get:
      tags:
        - Characters
      summary: Elenca tutti i personaggi della sessione per il DM
      description: Include anche personaggi OWNER_ONLY o DM_ONLY.
      security:
        - DmSessionBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Personaggi della sessione ordinati per creazione
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CharacterResponse"
        "401":
          description: Credenziale DM non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Sessione inesistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/dm/sessions/{sessionId}/close:
    post:
      tags:
        - Sessions
      summary: Conclude la sessione e disabilita il tavolo
      security:
        - DmSessionBearer: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Sessione conclusa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloseSessionResponse"
        "401":
          description: Credenziale DM non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    VenueKey:
      type: apiKey
      in: header
      name: X-BoardHub-Venue-Key
      description: Credenziale amministrativa del locale, non esposta ai QR o ai client pubblici.
    DmSessionBearer:
      type: http
      scheme: bearer
      bearerFormat: bhd1
      description: Credenziale del Dungeon Master, associata a una sola sessione attiva.
    PlayerBearer:
      type: http
      scheme: bearer
      bearerFormat: bhp1
      description: Credenziale HMAC associata a un partecipante e a una sola sessione attiva.
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      example: session-20260705-001
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateJoinRequestRequest:
      type: object
      properties:
        playerReference:
          type: string
          description: Identificativo locale stabile dell'installazione del giocatore.
          example: player-device-01
        displayName:
          type: string
          maxLength: 80
          example: Andrea
      required: [playerReference, displayName]
    JoinRequestResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        sessionId:
          type: string
        playerReference:
          type: string
        displayName:
          type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, EXPIRED]
        requestedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required: [requestId, sessionId, playerReference, displayName, status, requestedAt, expiresAt]
    ParticipantResponse:
      type: object
      properties:
        participantId:
          type: string
          format: uuid
        sessionId:
          type: string
        role:
          type: string
          enum: [DM, PLAYER]
        displayName:
          type: string
        status:
          type: string
          enum: [ACTIVE, LEFT]
        joinedAt:
          type: string
          format: date-time
      required: [participantId, sessionId, role, displayName, status, joinedAt]
    CreateCharacterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          example: Elaria
        species:
          type: string
          minLength: 1
          maxLength: 80
          example: Elfa
        age:
          type: integer
          minimum: 1
          nullable: true
          description: Età narrativa positiva; l'eventuale limite dipende dalla specie e dalla campagna.
          example: 120
        className:
          type: string
          minLength: 1
          maxLength: 80
          example: Maga
        level:
          type: integer
          minimum: 1
          maximum: 20
          example: 3
        speedCells:
          type: integer
          minimum: 0
          maximum: 100
          description: Velocità effettiva sulla griglia in segmenti da 5 piedi; il massimo è una protezione tecnica BoardHub.
          example: 6
        hpCurrent:
          type: integer
          minimum: 0
          nullable: true
          description: Se assente assume il valore di hpMax; non puo superarlo.
          example: 18
        hpMax:
          type: integer
          minimum: 1
          maximum: 1000000
          description: HP massimi effettivi; il massimo è una protezione tecnica BoardHub, non una regola D&D.
          example: 18
        armorClass:
          type: integer
          minimum: 0
          maximum: 100
          description: Classe Armatura effettiva; l'intervallo è una protezione tecnica BoardHub, non una regola D&D.
          example: 12
        partyVisibility:
          type: string
          enum: [OWNER_ONLY, PARTY, DM_ONLY]
          default: OWNER_ONLY
      required: [name, species, className, level, speedCells, hpMax, armorClass]
    CharacterResponse:
      type: object
      properties:
        characterId:
          type: string
          format: uuid
        sessionId:
          type: string
        participantId:
          type: string
          format: uuid
        name:
          type: string
        species:
          type: string
        age:
          type: integer
          nullable: true
        className:
          type: string
        level:
          type: integer
        speedCells:
          type: integer
        hpCurrent:
          type: integer
        hpMax:
          type: integer
        armorClass:
          type: integer
        partyVisibility:
          type: string
          enum: [OWNER_ONLY, PARTY, DM_ONLY]
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - characterId
        - sessionId
        - participantId
        - name
        - species
        - className
        - level
        - speedCells
        - hpCurrent
        - hpMax
        - armorClass
        - partyVisibility
        - version
        - createdAt
        - updatedAt
    AcceptJoinRequestResponse:
      type: object
      properties:
        request:
          $ref: "#/components/schemas/JoinRequestResponse"
        participant:
          $ref: "#/components/schemas/ParticipantResponse"
        accessToken:
          type: string
          description: Credenziale locale firmata da conservare sul dispositivo del giocatore.
          example: bhp1.550e8400-e29b-41d4-a716-446655440000.signature
      required: [request, participant, accessToken]
...
    CloseSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          example: session-20260705-001
        status:
          type: string
          enum: [ENDED]
        endedAt:
          type: string
          format: date-time
      required: [sessionId, status, endedAt]
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: SESSION_NOT_FOUND
        message:
          type: string
          example: "Sessione non trovata: session-non-esistente"
      required:
        - code
        - message
...